requirements.txt
fastapi
uvicorn
starlette
fastapi-sessions
mysql-connector-python